add --use-feature=class, part 1
This version adds option --use-feature='class' to format code using "use feature 'class'" (see rt145706).  The main changes are:

'class' is treated as a generalization of 'package'. It is updated to accept
attributes and an optional BLOCK.

'method' is handled using the existing --sub-alias-list option,
  i.e. --sub-alias-list=method

'field' is treated the same as 'my'

'ADJUST' is treated like 'BEGIN'

In a future update, the new option --use-feature="class" will be set
to be the default.  That requires some additional updates to avoid
conflicts with older uses of these keywords.