Fix MSWin32 EOL translation without UTF-8 in backup-mode
On Windows machines, the CRLF translation layer was left in-place, when using
the backup (-b) without the utf8-ecoding (-enc=utf8) option, which causes
the output file to be converted to "DOS" CRLF line endings, even with
the -ole=unix setting enabled.

Added `binmode` call in `backup_method_move` and `backup_method_copy` to
disable the CRLF translation layer, in the case of none-encoded-data, as
it is already used in `PATH 3b: output to a named file`.